DatedDurationShareDto as inner class of OvertimeAbsenceDto
